Restructure code and change some variable names.
