+++ app.js
@@ -1,5 +1,7 @@
-const { showMainIfActiveFast } = require('./cli-ui/cli-ui');
+const CliSingleton = require('./cli/cli');
 const chalk = require('chalk');
+
+const cliInstance = new CliSingleton().instance;
 
 console.log(chalk.rgb(147,96,176)(`
 -----------------------------------    
@@ -7,5 +9,5 @@ Welcome to the Zero fasting CLI app
 ___________________________________
 `));
 
-//Start active fast main menu initially
-showMainIfActiveFast();
+//Start the CLI
+cliInstance.startUi()
