+++ cli/cli-menu-structures/active-fast-main-menu.js
