+++ cli/cli-menu-structures/current-fast-info-menu.js
