+++ cli/cli-menu-structures/fast-types-options-menu.js
