+++ cli/cli-menu-structures/inactive-fast-main-menu.js
