+++ cli/cli-menu-structures/no-current-active-fast-menu.js
