+++ cli/cli-menu-structures/show-all-previous-active-fasts.js
