+++ cli/helper.js
