+++ data/data.json
@@ -1,1 +1,1 @@
-{"userData":{"allFastSessions":[{"_status":false,"_started":"2022-03-29T12:15:00.000Z","_ending":"2022-03-30T06:15:00.000Z","_type":"18"}],"currentFast":{"_status":true,"_started":"2022-03-29T18:20:00.000Z","_ending":"2022-03-31T06:20:00.000Z","_type":"36"}}}
+{"userData":{"allFastSessions":[{"_status":false,"_started":"2022-03-29T12:15:00.000Z","_ending":"2022-03-30T06:15:00.000Z","_type":"18"},{"_status":false,"_started":"2022-03-30T07:00:00.000Z","_ending":"2022-03-31T01:00:00.000Z","_type":"18"},{"_status":false,"_started":"2022-03-30T07:00:00.000Z","_ending":"2022-03-31T01:00:00.000Z","_type":"18"}],"currentFast":{"_status":true,"_started":"2022-03-30T07:05:00.000Z","_ending":"2022-03-31T01:05:00.000Z","_type":"18"}}}
